JSKA-5 Merged: JSKA-9 and JSKA-10 into Epic Story JSKA-5 with comments.
This code works with the sample input file in creating separate directories
for each mnemonic (JSKA-9) and creating values.h5 files inside each
of those directories (JSKA-10). Issues for creating times.h5 and
index.h5 are still in development in different branches.

This code is committed to a test/ branch until a review.